Refactor TTL headers to proper folders
