+++ src/ttl/include/ttl_cache/ttl_fifo_cache.hpp
@@ -6,7 +6,7 @@
 #include <optional>
 #include <unordered_map>
 
-#include "cache_statistics.hpp"
+#include "cache_statistics/cache_statistics.hpp"
 #include "math/saturation_arithmetic.h"
 
 class TTLFIFOCache {
