+++ src/ttl/include/ttl_cache/ttl_lru_cache.hpp
@@ -5,7 +5,7 @@
 #include <map>
 #include <unordered_map>
 
-#include "cache_statistics.hpp"
+#include "cache_statistics/cache_statistics.hpp"
 #include "math/saturation_arithmetic.h"
 
 struct myTTLForLRU {
